requirements.txt
fastapi>=0.118.3
uvicorn[standard]>=0.37.0
python-dotenv>=1.1.1
crawl4ai>=0.4.0
google-genai>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0